Consolidate garbage collection tests.
